user/requirements.txt
flask==2.3.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0
flask-bcrypt==1.0.1
requests==2.31.0